requirements.txt
accelerate==0.23.0
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
docker-pycreds==0.4.0
faiss-cpu==1.8.0
filelock==3.9.0
fsspec==2023.4.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.26.1
packaging==23.2
protobuf==4.25.3
psutil==5.9.6
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sentencepiece==0.1.99
sentry-sdk==1.43.0
setproctitle==1.3.3
shortuuid==1.0.12
six==1.16.0
smmap==5.0.1
sympy==1.12
tokenizers==0.14.1
torch==2.1.0+cu118
tqdm==4.66.1
transformers==4.35.0.dev0
triton==2.1.0
typing_extensions==4.4.0
ujson==5.9.0
urllib3==2.0.7
wandb==0.16.4
